refactor(useDeleteLayer): 요소 삭제 로직 useDeleteLayer hook으로 분리
